+++ src/providers/IntentManagerProvider.ts
@@ -78,7 +78,7 @@ export class IntentManagerProvider implements vscode.FileSystemProvider, vscode.
 	};
 
 	public onDidChangeFileDecorations: vscode.Event<vscode.Uri | vscode.Uri[] | undefined>;
-    private _eventEmiter: vscode.EventEmitter<vscode.Uri | vscode.Uri[]>;
+    private _eventEmiter: vscode.EventEmitter<vscode.Uri | vscode.Uri[]>; // = new vscode.EventEmitter(); it is defined in constructor
 
 	/**
 	 * Create IntentManagerProvider
@@ -193,10 +193,8 @@ export class IntentManagerProvider implements vscode.FileSystemProvider, vscode.
 						this._getNSPversion();
                     } else {
 						this.pluginLogs.warn("NSP response:", response.status, json.error);
-
 						DECORATION_DISCONNECTED.tooltip = "Authentication failure (user:"+this.username+", error:"+json.error+")!";
 						this._eventEmiter.fire(vscode.Uri.parse('im:/'));
-
 						this.authToken = undefined; // Reset authToken on error
                         reject("Authentication Error!");
 					}
@@ -218,8 +216,7 @@ export class IntentManagerProvider implements vscode.FileSystemProvider, vscode.
 
 	/**
 	 * Gracefully revoke NSP auth-token.
-	 */	
-
+	 */
 	private async _revokeAuthToken(): Promise<void> {
 		if (this.authToken) {
 			const token = await this.authToken;
@@ -266,7 +263,7 @@ export class IntentManagerProvider implements vscode.FileSystemProvider, vscode.
 					"Accept": "application/yang-data+json",
 					"Authorization": "Bearer " + token
 				};
-			else
+			else 
 				options.headers = {
 					'Content-Type': "application/json",
 					'Accept': "application/json",	
@@ -1421,7 +1418,6 @@ export class IntentManagerProvider implements vscode.FileSystemProvider, vscode.
 
 			DECORATION_CONNECTED.tooltip = "Connecting..."; // revert to original text
 			DECORATION_DISCONNECTED.tooltip = "Not connected!"; // revert to original text
-
 			this._eventEmiter.fire(vscode.Uri.parse('im:/'));
 		}